+++ src/plonemeeting/restapi/__init__.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 
 from zope.i18nmessageid import MessageFactory
+
 import logging
 
 
-_ = MessageFactory('plonemeeting.restapi')
-logger = logging.getLogger('plonemeeting.restapi')
+_ = MessageFactory("plonemeeting.restapi")
+logger = logging.getLogger("plonemeeting.restapi")