+++ src/plonemeeting/restapi/interfaces.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
+from imio.restapi.interfaces import IImioRestapiLayer
 from plone.restapi.interfaces import IPloneRestapiLayer
-from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 
-class IPMRestapiLayer(IPloneRestapiLayer):
+class IPMRestapiLayer(IImioRestapiLayer, IPloneRestapiLayer):
     """Marker interface that defines a browser layer."""