+++ src/plonemeeting/restapi/serializer/meeting.py
@@ -11,14 +11,10 @@ from zope.interface import Interface
 @implementer(ISerializeToJson)
 @adapter(IMeeting, Interface)
 class SerializeToJson(BaseATSerializeToJson):
-    ''' '''
+    """ """
 
-    def __call__(self, version=None, include_items=True):
-        ''' '''
-
-        result = super(SerializeToJson, self).__call__(
-            version=version, include_items=include_items)
-
+    def _additional_values(self, result):
+        """ """
         # add some formatted values
-        result['formatted_assembly'] = self.context.displayStrikedAssembly()
+        result["formatted_assembly"] = self.context.displayStrikedAssembly()
         return result