+++ src/plonemeeting/restapi/setuphandlers.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+
 def post_install(context):
     """Post install script"""
     # Do something at the end of the installation of this package.